fix: redirect directly to edit template page
