+++ apps/journeys-admin/src/components/TemplateView/TemplateViewHeader/TemplateEditButton/TemplateEditButton.tsx
@@ -15,7 +15,7 @@ export function TemplateEditButton({
   const { t } = useTranslation()
 
   return (
-    <NextLink href={`/publisher/${journeyId}/edit`} passHref legacyBehavior>
+    <NextLink href={`/publisher/${journeyId}`} passHref legacyBehavior>
       <Button startIcon={<Edit2Icon />} data-testid="TemplateEditButton">
         {t('Edit')}
       </Button>
